requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
pydantic
faiss-cpu
pandas
numpy
python-dotenv